Added additional build dependencies to README
Added the following depenencies to the README file:
  * flex
  * autoconf
  * automake
  * libtool
  * uuid-dev
These are based on what additional packages needed to be installed over a clean
Ubuntu 11.10 installation.

Changed the requiment for a C99 compiler to "gcc/g++". The unit tests require a
c++ compiler, and it is better to recommend a specific compiler that works. We
can add others if they are known to work.

Signed-off-by: Michael Schonberg <mschonberg@nyx.com>