Move Linux Specific #includes and Macros
This is the first in a series of patches to add Windows suport. Rather
than peppering the source with #ifdef osXXX blocks, os specific headers
and source files go into directory with the same name as the OS. Ideally,
there will be one linux directory in common, one windows, etc.; however,
the same approach should be used in other modules that require os specific
implementations. "#ifdef OS" should be avoided whenever possible.

Each OS directory must at least contain a file "port.h". This file
contains os-specific includes and macros. At some point, it might make
sense to subdivide port.h into several headers (one for networking,
threads, etc.). For the moement we will stick with one until the process
is complete.

We may also want to move some of the linux/ that is common to other os's
like Solaris and BSD to a /unix directory when we start adding support for
other unix-like platforms.

This change creates common/c_cpp/src/c/linux/port.h. All of the #includes
that are not found by the windows compiler are in port.h as well as some
of the obvious linux specific macros.