Added WCACHEExpDLL macro for building NYSE wirecache
This is all part of DLL hell for windows: when we build common we need to
indicate that the mehtods from common are export from the DLL, and then we need
to indicate that they are imported when we build other products. This is
further complicated when products mix static and dynamic linking for different
components.