Preparing timer.c and timer.c for windows support
These files are required by the Avis bridge as Avis does not provide
native timer support. The changes include changing pthread_xxx calls to
wthread_xxx calls, and implementing macros for socketpair and setting
nonblocking because windows does not support socketpair() and uses a
slightly differfent call for setting non-blocking on a socket.

Signed-off-by: Mike Schonberg <mschonberg@nyx.com>