Added wthread_global_mutex macros
In preparation for windows support, replace global phread_mutex_t and
wthread_mutex_t inititialized with PTHREAD_MUTEX_INITIALIZER with
wthread_gobal_mtuex() macros. This is required because Windows critical
sections and mutexes do not have static initializers. They need to be
protected with atomic access to a global state variable.

The macros are simple for Linux:
	WTHREAD_GLOBAL_MUTEX(NAME) initializes the mutex
	wthread_mutex_(un)lock(name) locks and unlocks.

For Windows, the lock macro increments a global reference count to ensure
that the mutex only gets initialized once, and spins on a another variable
to wait until initialization complets. See windows/port.h (future check
in) for details.

Signed-off-by: Mike Schonberg <mschonberg@nyx.com>