Ownership of message buffers
The msg payload bridges now have a msgPayloadSetParent() call which is
used  to set the parent mamaMsg object for a payload. This  allows the
"ownership" of the mamaMsg to be interrogated from a payload bridge
i.e. if a message  can be modified or not. The use of the message owner
term has also been normalised.

Signed-off-by: Glenn McClements <gmcclements@nyx.com>
Signed-off-by: Mike Schonberg <mschonberg@nyx.com>