[mamda] map a delete msg through mamda as an error callback
When using a MamdaSubscription and a msg of type delete is received, Mamda
handled this internally and did not notify the user. As no more messages are
received after the delete msg this symbol effectively stopped ticking.  This
patch fires a callback on the MamdaErrorListener to inform the user of the
delete msg so they can take any action they wish based on this information. (eg
delete the subscription.)

Signed-off-by: Ian Bell <IBell@nyx.com>