MAMAC TESTTOOLS: Additional logging to indicate processing in capture replay
When reading the playback file during startup, capture replay may sit for several
minutes (when reading a large file), without any indication that it is processing
data. This patch adds some additional logging to make it clear that the file
continues to be processed, as well as giving a clearer indication as to the number
of symbols discovered.

Signed-off-by: Damian Maguire <dmaguire@nyx.com>