UNITTEST: Several fixes for msggeneraltests tests and return codes
Detaching depends on a bridge and queue being initialized - these are
now in place for msgDetachValid.

Disabled the msgImplSetMessageOwnerInValidOwner test because from what
I understand, what is provided is actually a valid owner.

Fixed msgImplSetQueueValid to use a valid queue.

Modified msgImplGetStatusFromMsgValid and msgGetEntitleCodeValid to
operate on a conforming MAMA message which contain required fields
field.

Fixed msgGetDateTimeMSecValid, msgGetFieldAsStringValid and
msgSetNewBufferValid which was calling underlying functions incorrectly.

Signed-off-by: Frank Quinn <fquinn.ni@gmail.com>