MAMDA: Iterator bug in MamdaOptionChain.cpp
Resolve a bug in MAMDA Option Chain function getStrikesWithinRangePercent,
where the iterator is never incremented and the code falls into an
infinite loop. A unit test should be written to check the behaviour of
this code.

Note: This bug also causes issues for optimised Windows builds - cl.exe can
fail to compile with a Fatal Error (C1001), as it cannot generate code for
the construct. Changing to a debug build, or removing optimisation allowed
the build to continue.

Signed-off-by: Damian Maguire <dmaguire@nyx.com>