MAMA: Completed mamaDateTime hints implementation
The calls to mamaDateTime get / set with hints calls provide parameters for
hints but they do not seem to be wired. When you look at the bit masks for
seconds, micros, precision and date time presence, there is a spare byte
there which I have used to store hints in this patch. Verified in unit
test MamaDateTimeTestC.TestGetWithHints.

Signed-off-by: Frank Quinn <fquinn.ni@gmail.com>