Subject: [PATCH 4/5] UNITTESTS: Fixed the payload vector time / price unit tests
Several of the unit tests for payloads were passing incorrect parameters
and were evaluating the results incorrectly. This patch resolves this as
well as re-enabling some old unit tests which seem to be valid now that
we have a payload implementation that supports them. Casting for
compatibility with the existing interface was also centralized with
macros.

Signed-off-by: Frank Quinn <fquinn.ni@gmail.com>

Later modified as previous patch in series will not be committed until
major version released. As a result some tests should not yet be enabled.
General changes tidying up the tests and varifying correct behaviour have
been left in.

Signed-off-by: Matt Mulhern <mmulhern@nyx.com>