Subject: [PATCH 2/5] UNITTEST: Fixes for MAMA Vector Time / Price Unit Tests
Corrected some issues around expected behaviour and centralized the casts
for the inconsistent MAMA interface around vector price / date times
so that it can be easily removed if / when it eventually gets corrected.

Signed-off-by: Frank Quinn <fquinn.ni@gmail.com>

Later modified as previous patch in series will not be committed until
major version released. As a result some tests should not yet be enabled.
General changes tidying up the tests and varifying correct behaviour have
been left in.

Signed-off-by: Matt Mulhern <mmulhern@nyx.com>