Revert "MAMDA EXAMPLES: Add support to bookviewer for non-MAMAIO middlewares"
This reverts commit 1eb8eaeb2789134e6e4ef10c3c2e593a4dbadfc7.

The original implementation here was to provide the ability for the
bookviewer application to run against the Qpid Proton reference bridge.
This is no longer required (since the bridge now has it's own IO
implementation), and the change doesn't really fit the requirements
of the example app. Reverting this change until such a solution is
implemented.