COMMON: Fix issue with memsetting static const variable.
Currently the gIPAddress in common/linux/network.c is declared
as static const.  This is causing a run time error when compiled
with Clang not seen on GCC - reproducible with mamalistenc.  The
error seen is a "Bus Error: 10" on Darwin and "Segmentation Fault"
on Linux.  Extract from back trace as follows:

Program received signal SIGSEGV, Segmentation fault.
lookupIPAddress () at common/c_cpp/src/c/network.c:166
166         memset( (void *)gIPAddress, 0, 16 );
(gdb) bt
    subscMsgType=<error reading variable: Unhandled dwarf expression opcode 0x0>, msg=<optimized out>, issueSymbol=<optimized out>)
        at mama/c_cpp/src/c/msgutils.c:286
        ...

The issue is that with Clang is placing gIPAddress in read-only memory,
and we get memory issues when we call memset on that location.  The
change is to remove the const qualifier.

Tested again with mamalistenc and behaving correctly and using debugger
- can see the IP address getting populated correctly:

        gIPAddress      char [16]       "10.211.55.2"

Signed off by Philip Preston <philippreston@mac.com>