COMMON: Portability fixes for Common queue size and MUTEX code
Changes that improves OS portability, and will allow further
commits to support MAC OS X. The changes are for existing
operating systems and do not affect existing functionality.

There are two main changes:

1. PTHREAD_MUTEX_RECURSIVE_NP

Currently when setting pthread recursive type we use the “NP”
type which is “Non Portable”. This is not available on all
operating systems. The change is to hide the pthread type
behind a macros (WTHREAD_MUTEX_RECURSIVE), which will be set
in port.h for each operating system. The rest of the code base
will use this macros for initialising the thread.

2. Queue Size

Currently the Max queue size for common is set in queue.h, where
a macros checks for the presence of SEM_VALUE_MAX. If not available
its set to a hard coded value. This makes sense to move to port.h
as whether it uses SEM_VALUE_MAX or hard coded value is operating
system dependant. It will also allow setting of queue sizes when
SEM_VALUE_MAX is not an appropriate size

Tested on:

* Mac OS X 10.9
* Ubuntu 12.04 x86_64
* Windows Server 2008

All unit tests passing.

Signed-off-by: Phil Preston <philippreston@mac.com>