Updating jobs Command Line Implimentation
Summary:
Previous implementation utilised the scons built in parameter. Adding a new
argument called jobs, which will accepts 1...n as the number of jobs to spawn
where n is all available cores. This will be saved in the conf file and so
will be set for all subsequent rebuilds

Test Plan:
Run scons build with various versions of the jobs parameter, and
ensure it runs over appropriate numbers of cores.

Reviewers: frank

Reviewed By: frank

Differential Revision: http://review.openmama.org/D9