QPID: Moving meta info to application properties
Currently we add a number of custom properties to the properties section of the AMQP message (e.g. MsgType, Inbox Name).  The properties section of the message is used for a defined set of standard properties of the AMQP spec (message-id, user-id etc), and is not meant to be extensible for custom properties.  The Application Properties section of the AMQP message is the section for adding custom properties, in the form of a key (string), value map.

This change moves the following information from properties to application properties:

* Message Type
* Inbox Name
* Reply To
* Target Subject

Tested for:

* Pub / Sub
* Pub / Inbox
* CaptureReplay / MamaListen

Signed-off-by: Phil Preston <philippreston@mac.com>