QPIDMSG: Fixed issue with reallocating vector fields
This patch fixes an issue where a realloc occurs for member types which the
qpid bridge will typically attempt to re-use rather than create / destroy
every time. This is really just MAMA Message at this time but the same
would apply to date time and price if they were implemented. Without this
change, if the newly appended memory happens to be non-NULL, the qpid bridge
will likely crash.

Signed-off-by: Frank Quinn <fquinn@nyx.com>