MAMA: Moved file parsing to external module for playbacks
Summary:
This change moves file parsing capabilities to an external module. This
makes the code more readable and modular as well as allowing us to remove
the file streams macro which was temporarily added to allow us to run the
playback tools on windows.

The rightCopy function was also removed as a copy was not actually
required, and mamaPlaybackFileParser_setSize was also removed in favour of
the new fileParser functions.

There is also a new configuration parameter to decide whether to use MMAP
or file streams. Mmap is now the default for all systems, but some users
may prefer to use file streams instead due to its reduced memory
footprint. To enable file streams, the following configuration parameter
may be used:

mama.playbackfileparser.use_mmap=false (default = true)

This was tested on both Linux and Windows and included testing on
windows and linux for both mmap and file stream modes using
captureconvertc and capturereplayc with a mamalistenc client.

Signed-off-by: Frank Quinn <fquinn@nyx.com>

Test Plan: Check complete build and unit tests. Check capturereplay variants.

Reviewers: frank

Reviewed By: frank

Differential Revision: http://review.openmama.org/D21