COMMON: Fix race condition in common-based queue unblocking.
Summary:
The queue in common code now uses an atomic (wInterlockedInt) variable
for tracking whether or not the queue is being unblocked (which is often
as a result of requiring to shut the queue down).

A race condition was observed when calling wombatQueue_unblock() in
which it was possible for the semaphore to be posted and the dispatch
thread to resume before the unblocking variable was set. This was seen
in helgrind, valgrind and in practice (middleware queue hanging).

This would result in a dispatch happening (regardless of whether there
is an item or not) and the dispatch thread continues to block. Prior
to the empty queue fix made in commit c985f942 this would have caused a
crash if the queue was empty.

The fix from commit c985f942 has been left in but it could probably be
taken out now that the unblocking variable is protected (we shouldn't
ever get into the body of dispatch now without an item actually being
available to dispatch).

No other changes - CommonQueueTestC.* unit tests executed successfully.

Signed-off-by: Lee Skillen <lskillen@vulcanft.com>

Test Plan: Check build and common queue unit tests.

Reviewers: frank

Reviewed By: frank

Differential Revision: http://review.openmama.org/D5