SCONS: Resolve issues with scan-build tool.
Summary:
There are a couple of issues with the current scan-build tool
within our scons framework, where it doesn't pass the full set of
environment variables to the underlying compiler (in this case
ccc-analyze and cxx-analyze).

In order to resolve, the tool now sets the SCons env['ENV']
variables, which forces them to be passed to the environment
under which the analyzers run.

Also added some additional logic which ensures that the
clang-analyzer cannot be invoked except when run within scan-build,
and to fix issues with compiler specification when running the clean
command (clean should proceed regardless of the compiler specified).

Signed-off-by: Damian Maguire <damian@sciantechnologies.com>

Test Plan: Checking building OpenMAMA with scan-build and scons.

Reviewers: frank

Reviewed By: frank

Differential Revision: http://review.openmama.org/D27