UNITTESTS: Adding getFieldAsString buffer size tests
Summary:
We've previously seen issues with getting vector messages as strings
when the buffer size is too small for the vector. These unit tests
exercise the ability of payloads to handle both cases where the buffer
is large enough, and when the buffer is too small to contain the
message.

Signed-off by: Damian Maguire <dmaguire@nyx.com>

Test Plan: Run unit tests.

Reviewers: frank

Reviewed By: frank

Differential Revision: http://review.openmama.org/D23