COMMON: Resolve issues with closing sockets on Windows
Windows sockets cannot be closed using the same 'close' call as Linux
sockets. As such, when we hit the previous calls to close, it causes
issues on Windows machines. This change swaps 'close' for a
'wclosesocket' call, which is set to platform specific calls within
the port.h headers of each platform.

Signed-off by: Damian Maguire <dmaguire@nyx.com>