MAMA: Update to the SyncResponder to prevent free'd memory being read.
Summary:
When a sync request is received symbol names are copied retrieved from
the transport list. However this is a not a deep copy and if a subscription
is removed from the list before the response is sent, the response can be
corrupted. Taking a deep copy of the symbol name(s) will prevent this.

This is a subtle race condition which could not be recreated directly
with a MAMA application. We were able to recreate it consistently with an
enterprise application and can confirm that the issue is resolved.

OMAMA-271

Signed-off-by: Gary Molloy <gmolloy@nyx.com>

Test Plan: Investigate code changes to see if any other tests can be performed.

Reviewers: frank

Reviewed By: frank

Differential Revision: http://review.openmama.org/D37