COMMON: Add check for error return code from read
When reading the stat file in getProcessInfo, it is possible for the
file to be removed and the read to fail. If we don't check for the -1
return, we can end up clobbering a bit of the stack, and hitting other
problems later.

Signed-off by: Damian Maguire <dmaguire@nyx.com>