MAMA: Made logging more accurate in MAMA Ft
In some MAMA FT Timers, the timer callback always reports that it is a
"multicast" FT type, when in fact that code is shared for all FT types.
This is misleading to the MAMA application developer who is trying to
troubleshoot FT issues, so this patch will make the logging a little
more clear.

Signed-off-by: Frank Quinn <fquinn@nyx.com>