RPM: Addition of the Spec file for building OpenMAMA RPMS
This change adds the spec file for generation of the OpenMAMA release
RPMs, under a new top level directory 'release'. The aim is over time to
provide a full suite of scripts which can be used to generate an
OpenMAMA release across multiple platforms.

The Spec file itself is reasonably straight forward, and generally
avoids any particularly complex constructs.

However it has been designed with a very specific build environment in mind,
and may not work out of the box. This is due to, among other things, the
requirement for a specifically packaged .tgz which includes sample data
files, and the relocation of files from within the source to alternative
locations.

In a later commit we will add a build script which should make it possible
for users on supported platforms to build out of the box.

At present the spec file has been tested on RedHat and Centos 6, and
Fedora 19 and 20. If you use any other platforms YMMV.

Signed-off by: Damian Maguire <d.maguire@srtechlabs.com>