RPM: Addition of RPM release builder script and helper functions.
The openmama-rpm.sh script is used to generate a full OpenMAMA linux
release, including multi-platform RPMs, the SRPM, a source tarball and
an x86_64 binary release.

For most uses the syntax is straight forward. Simply executing the
script will, within the current directory, checkout a fresh copy of next (1
commit deep); checkout the sample data; prepare a source tarball; build an
RPM using the rpmbuild tool; create RPMs for other (currently hardcoded)
platforms using the mock tool; and package up everything within a
'release'.

Sample output of a successful invocation:

    $ ./openmama-rpm.sh
    Setup build environment.                                   [  OK  ]
    Checking out clean source tree                             [  OK  ]
    Checking out the OpenMAMA sample data                      [  OK  ]
    Create the source tarball                                  [  OK  ]
    Perform the RPM Build                                      [  OK  ]
    Build multiplatform RPMs                                   [  OK  ]
    Packaging release                                          [  OK  ]

There are a number of other command line options which may be invoked.
Full help text is available by executing the script with the --help
argument (or -h|-?). However, basic usage is as follows:

usage: openmama-rpm.sh [-n|--next] [-v|--version <version>]
                       [-b|--build-number <build-number>]
                       [-B|--build-dir <build directory>]
                       [-d|--data-dir  <data directory>]
                       [-s|--source-dir <source directory>]
                       [-r|--release-dir <release directory>]
                       [--env-only] [--clone-src-only] [--clone-data-only]
                       [--tar-source-only] [--rpm-only] [--mock-only]
                       [--package-only] [-h|-?|--help]

As with the spec file, the script is *not* general purpose, and does
make some assumptions about the setup of the current environment. This
includes the fact that you have rpmbuild and mock installed, that you
are a member of the mock group, that you are not running as root
(seriously, we'll try and drop out if you are, so don't), that you have
an active internet connection (so we can grab the OpenMAMA source), and
that you have all the pre-requisites for an OpenMAMA build in place (in
the default locations).

It should work on RHEL and Centos 6, Fedora 19 and 20. I'd expect it to
also work on RHEL/Centos 7, and Fedora 21, though tweaks may be
required.

Signed-off by: Damian Maguire <d.maguire@srtechlabs.com>