No value expansion of last property line in mama.properties
On load every property inside the mama.properties is expanded if
environment variables exist. For example:

In reality all properties are expanded if they have new line after them.
If a property happens to be on the last line in mama.properties, then it
won't be loaded and expanded as result of that.

BZ-156

Signed-off-by: Guy Tal <guy.tal@tick42.com>