Problem with mamaDictionary_getDictionaryMessage when multiple bridges are loaded
Currently mamaDictionary_getDictionaryMessage() will return you a
message using the default payload which may not be the payload type for
the bridge that is trying to populate a message with its dictionary.

This patch will provide an alternative function to populate a mamaMsg
with the dictionary. This differs from the existing function by allowing
the caller pass a mamaMsg that has been pre-created on whatever payload is desired.

BZ-178

Signed-off-by: Tom Doust <tom.doust@tick42.com>