UNITTEST: Fixed assumption of order in iteration
The unit tests assumed that the order of entry into a message
field will be identical to the order of extraction while iterating.
This proved not to be the case with avis and possibly other payload
implementations so these changes will remove that assumption
without losing out on any other existing test strictness.

Signed-off-by: Frank Quinn <fquinn.ni@gmail.com>