UNITTEST: Remove presumptuous isTportDisconnected test
I don't think it's reasonable to expect a unit test around OpenMAMA
to support this function call as the nature of transport connectivity
will vary across all middlewares. In the instance of Avis, it is
never really in a disconnected state and there are doubtless other
middleware implementations that are not connection oriented at all
and for which this test doesn't really make any sense, so I'm
taking it out.

Signed-off-by: Frank Quinn <fquinn.ni@gmail.com>