QPID: Fixed issue with qpid deadlock in timer unit tests
This bug was introduced recently when trimming out some shutdown
code which was deemed to no longer be required. However it did
put the application in a state where a ctrl+c was required to
interrupt the blocking receive loop. This change introduces a
timeout of 1ms (non-configurable at the moment) for each recv
call which allows the thread to avoid deadlock.

Signed-off-by: Frank Quinn <fquinn.ni@gmail.com>