MAMA: Remove dependency on mamaMiddleware enumeration
Summary:
This patch removes the dependency on the mamaMiddleware enum for loading
and interacting with middleware bridges.

It replaces the mamaImpl structure's myBridges array with a wtable,
keyed off the middleware's string name. Rather than storing the LIB_HANDLE and
mamaBridge structures separately, it also combines the two into a new
mamaBridgeLib structure, which are stored in the wtable.

Locations where we formerly accessed or iterated over the gImpl.myBridges array
have had their functionality replaced with wtable versions. In most instances
the aim has been to replicate the existing functionality, or as close to
as possible.

We've also deprecated the use of the mamaInternal_registerBridge
function, which is no longer required, and is now a no-op. The function
has also been removed from the INITIALIZE_BRIDGE macro, so it is no
longer called.

Finally, we've added a mamaInternal_init function, which is used to
initialise any structures used by OpenMAMA. At present this is only
called during bridge loading, however it may be expanded later.

MAMA: Remove dependency on mamaPayloadType enumeration

This patch removes the dependency on the mamaPayloadType enum for
loading and interacting with payload bridges.

It firstly replaces the mamaImpl structure's myPayloads array with a
wtable keyed off the payloads string name. Rather than storing the
LIB_HANDLE and mamaPayloadBridge structures seperately, it also combines
the two into a new mamaPayloadLib structure, which is stored in the
mPayloads table.

We have also implemented a mPayloadCache, which is a simple array
indexed by char identifier. This allows easy access for legacy features
which rely on the char identifier to the underlying payloadBridge, while
preserving flexibility.

Locations where we formerly accessed or iterated over the
gImpl.myPayloads array have had their functionality replaced with either
a wtable based implementation, or a mPayloadCache based implementation,
as appropriate. In most instances, the aim has been to replicate the
existing functionality, or as close to it as possible.

We've also cleaned up some middleware tests which were causing issues
with the stricter checking on payload loading which was part of the main
chunk of this patch.

Test Plan: Check everything.

Reviewers: frank, matt, glenn

Differential Revision: http://review.openmama.org/D31