MAMA: Enable library search based vtable population.
The current mechanism for managing population of the middleware and
payload vtables (within the mamaBridge and mamaPayloadBridge structs) is
to allow the bridges themselves to perform the loading. This creates
numerous problems, particularly when it comes to changing the interface,
or extending it.

This patch adds the ability for MAMA to populate that vtable itself, by
performing a search through the loaded bridge library for the functions
required. This will allow further expansion in the future, including
providing support for default implementations, and other features.