COMMON: Fix dateTime issue w/ Windows putenv in SetDate
putenv works differently on Windows - if the value is empty it removes the key from the env.
On Linux it sets the value as empty.
When gmtime is called it works differently if the TZ var is empty or missing.
gmtime will default to UTC if the TZ="", and to local TZ is missing.
Solution is to set TZ to UTC rather than empty in port.c.
Also add unit test for this bug in mama/c_cpp/src/gunittest/cpp/MamaDateTimeTest.cpp.

Signed-off-by: Reed Alpert reed.alpert@gmail.com