COMMON: Fix ret code from Win close shared lib
The Win sytem call to close a shared lib returns 1=ok and 0=error.
The mama call does the opposite, so invert the Win return code.
The only current effect is an invalid error msg when unloading plugins on Win.
No unit test as we need a 'dummy' plugin to load/unload.

Signed-off-by: Reed Alpert <reed.alpert@gmail.com>