Dictionary: add fid and name to msg for unknown type
The dictionary message for unknow field type changed to include fid and name.

In populateMessageFromDictionary() if the dict has an unknown type the field
is discarded but the msg only says it was an unknown type. The patch will add
the fid and name to the log msg to help locate the field. Includes new unit
test.

Also adjust indentation.

Signed-off-by: Reed Alpert <reed.alpert@gmail.com>