MAMA: Fixed issue with Payload ID not working with messages
The payload bridge now contains a back-reference to the mamaPayloadLib.
The mamaPayloadLib contains meta data about the payload such as which
payload ID was decided on for representing the payload on the wire.

Without this change, any calls from bridges to mamaMsgImpl_getPayloadId
will return NULL unless the message in question has been created from
a byte buffer.

Signed-off-by: Frank Quinn <fquinn.ni@gmail.com>