COMMON: Fixed race condition in timer creation / dequeueing
There was a race condition as per issue 109 where it was possible to
dequeue a timer element before its timer heap pointer had been
populated in createTimer (the one in timers.c).

This fix populates this pointer before queueing and on failure will
reset its value to NULL before returning as well as returning an
error code.

Signed-off-by: Frank Quinn <fquinn.ni@gmail.com>