Staged merge of Reed's changes for https://github.com/OpenMAMA/OpenMAMA/pull/100/files
This is a staging branch which will once again get squash merged to next when it lands.

Something happened to squash the commit history on reed's original branch to include
some of those fixes along with every change which had been submitted on next since the
original fork was taken. In addition, when going through some of these changes, some
changes which weren't on the original pull request have managed to creep in from
somewhere.

Raising this with a mind to create a fresh pull request which will merge safely with
the next branch and with *only the required changes* included.

Signed-off-by: Frank Quinn <fquinn.ni@gmail.com>