QPID: Added some disconnection handling to qpid.
It's a bit crude because qpid doesn't seem to support callback based
event handling but it'll do the job. It's a 3-strikes system which is
specific to each publisher. If a link is down 3 times in a row, the
publisher's endpoint for the failing URL is removed so it will no
longer attempt to publish to that destination. There is still some
work to do around memory management of the endpoint as it's now a
more complex object.

It was debatable whether to adopt this approach or whether to just
crawl all endpoints for the provided URL. On consideration I thought
that would be both slow and very intolerant of reconnecting clients
so I stuck with the 3-strikes per publisher approach.

There is still some more work to do around

- Memory management of the endpoint now it's a more complex object
- Ensuring that endpoints are only created for p2p transports
- I spotted a crash possibly related to issue #153 when (after the
  third client fails), the first client then disconnecting will
  crash
- Valgrind run

Signed-off-by: Frank Quinn <fquinn.ni@gmail.com>