QPIDMSG: Added implementation for the vector price / time functions
This patch adds support for vector price / date time to the qpid payload
bridge. It's worth noting that the function prototypes have also been
updated to reflect the way in which these functions are actually
expected to behave. This makes the code easier to understand for third
party payload bridge developers, but has the unfortunate side effect
of producing warnings while compiling until the MAMA Msg interface
is updated to reflect the same function sets.

Signed-off-by: Frank Quinn <fquinn.ni@gmail.com>