QPID: Added some disconnection handling to qpid (#65) (#162)
It's a bit crude because qpid doesn't seem to support callback based
event handling but it'll do the job. It's a 3-strikes system which is
specific to each publisher. If a link is down 3 times in a row, the
publisher's endpoint for the failing URL is removed so it will no
longer attempt to publish to that destination. There is still some
work to do around memory management of the endpoint as it's now a
more complex object.

It was debatable whether to adopt this approach or whether to just
crawl all endpoints for the provided URL. On consideration I thought
that would be both slow and very intolerant of reconnecting clients
so I stuck with the 3-strikes per publisher approach.

Signed-off-by: Frank Quinn <fquinn.ni@gmail.com>