COMMON: Added support for format strings when evaluating properties
Added a new method properties_GetPropertyValueUsingFormatString to allow the
caller to provide a sprintf style format string inline and provide a default
to reduce the amount of code required when performing common property parsing
tasks.

Also simplified properties_GetPropertyValueAsBoolean to use strtobool instead
of duplicating the strings to check for boolean evaluation.

Signed-off-by: Frank Quinn <f.quinn@srtechlabs.com>