COMMON: Fixed leaking and corruption in properties parsing
There were a few places in properties where memory was being poorly managed.
This change should simplify a fair chunk of it by removing the requirement
for the properties implementation to maintain a reference of all keys which
were strdup'd.

I also fixed a memory leak in properties_merge where repeated merging of
properties objects with identical keys would result in value objects leaking.
This was exposed by an issue in MAMA where properties files would be
double-parsed, then merged with itself depending on order of startup, causing
a copy of each of the values to leak.

Signed-off-by: Frank Quinn <fquinn.ni@gmail.com>