CI: Added new script to use for CI builds
Added initial version of CI script to allow CI builds to be run offline.
Note that it depends on the caller having a standard directory structure
set up for dependencies which is obviously OK for a CI environment where
this will always be consistent, but the directory structure itself will
need to be documented and advertised as well.

Signed-off-by: Frank Quinn <fquinn.ni@gmail.com>