SCONS: Fixed build issue with multiple build types on windows (#195)
This fixes an issue where it was impossible to build multiple build types
using one command on windows. It also brings unit tests to the scons
windows builds.

Note that qpid_home on windows also points to the base proton directory
now rather than requiring the user to know to point to the 'proton-c'
subdirectory.

When building types beyond dynamic, this also flagged an issue where
several functions were not exported for DLL properly, particularly in
mamaFieldCache so this patch also resolves that issue by introducing the
correct function declarations.