COMMON: Wired up implementation of strptime
There was a definition for strptime but it turned out that this wasn't
defined in any header files. It also appears that strptime depends on
strncasecmp which would have needed to be defined via port.h so I'm not
quite sure how this would have worked before.

Wired up implementation via definition in port.h, added Solution file for
common unit tests and added unit test for strptime.

Signed-off-by: Frank Quinn <fquinn.ni@gmail.com>