MAMA: Fixed ability to use custom properties file
Without this change, there is no way to specify a non-default properties
file which will be honoured by parts of the MAMA loading process including
the bridge loading stage without reporting an error.

Added new function for counting the number of properties which have been
specified and for counting the number of elements in a wtable. This is to
enable effective checking of whether or not property loading has been
successful so it can report an error only when necessary.

It will now only report an error if:

 - A specific properties file was requested
 - Default properties are being loaded and no existing properties are loaded

To use custom properties, use mama_setPropertiesFromFile or mama_setProperty
before mama_loadBridge is called.

Signed-off-by: Frank Quinn <fquinn.ni@gmail.com>