MAMA: Added utility functions to read symbol file
Added the start of a new set of functions in mamautils.h to
provide standalone stateless functions that provide
miscellaneous tasks. The first one is mamaUtils_readSymbolFile
and exists to centralize functionality that is duplicated in
various programs and example applications.

The first use of this is in captureconvert.c which will now
allow a whitelist of symbols to be provided. If a parsed MAMA
message is for a topic not in this list, it will not be provided
in the output. It's a good way to reduce playback sizes, filter
out orderbooks / summary books etc.

Signed-off-by: Frank Quinn <fquinn.ni@gmail.com>