EXAMPLE: Added -rewind option to mamaproducerc_v2
This option allows users to specify a number of times to rewind a
playback file, or use -1 for infinite rewinds. Without this change,
mamaproducerc_v2 would always exit when a playback had completed
its first run through.

Also added the other playback parameters to the help text to shed
some light on how to use the replay functionality.

Note that unlike capturereplay, this tool does not support initial
and recap responses as it does not use the DQ Publisher.

Signed-off-by: Frank Quinn <fquinn.ni@gmail.com>