COMMON: Removed wtimegm implementation for Windows
The wtimegm implementation was previously a custom function
implemented by OpenMAMA in port.c for windows. This caused issues
when trying to run unit tests on Windows as it was triggering
random access violations from that area of code.

This change replaces this custom function with _mkgmtime on
windows which has been around since Visual Studio 2005:

https://msdn.microsoft.com/en-us/library/2093ets1(v=vs.140).aspx

After this change, the access violations on Visual Studio while
running unit tests have now stopped.

Signed-off-by: Frank Quinn <fquinn.ni@gmail.com>