Added timespec based implementation for datetimei and updated C# and JNI to use timespec. Code cleanup has been done too.
The timeval implementation works only for Linux based systems. Original
design documents assumed that the 'tv_sec' timeval on windows was a
time_t (as it is on linux), but it turns out that it is in fact a long
int, which is 4 bytes on windows for both 32 and 64 bit.

To provide a very similar alternative, this change introduces new
mamaDateTime methods which make use of timespec which is similar to
timeval, but uses time_t to store seconds and has nanosecond resolution.

This change also includes a definition of struct timespec for older MSVC
compilers. The macro HAVE_STRUCT_TIMESPEC may be also be provided to
explicitly disable this definition if a duplicate definition is already
declared in other headers.

We also updated time spec extern declarations, Java Unit tests to build
on Windows and JNI Date Time to use timespec structs now.

Signed-off-by: Marti Queralt <m.queralt@daxplat03v.belfast.wombatfs.com>