Modified OpenMAMA RPM script to use Mock Tarball
The CI boxes recently had the qpid proton rpm removed from them in
favour of a build which was not linked against OpenSSL to reduce
spurious valgrind warnings coming out of qpid proton. This broke
the RPM build though as it attempts to do an inline build of the
RPM on the host which expects dependant RPMs to already be
installed.

These changes instead generate the release tarball from an
existing mock environment, and restrict host rpm builds to simply
generating the source RPM.

The naming of snapshot tarballs has also been modified with a mind
to serving them up as artifacts in jenkins.

Signed-off-by: Frank Quinn <fquinn.ni@gmail.com>