Fixed issue with multiple subscribers for same topic in qpid
See issue #262. This was most likely introduced when memoryPool changes were
made, but there was an issue where a proton message was not initialized for
the temporary copies that are created when multiple subscribers are consuming
the same message. In addition there was another issue where the memory node
itself rather than its underlying buffer was being manipulated. This change
fixes both of the above.

Also fixed an issue with memoryPool_iterate which could cause all members
of the memory pool not to trigger a callback.

Signed-off-by: Frank Quinn <fquinn.ni@gmail.com>