Fixed issue with date time on 32 bit linux (#275)
Only 32 bit linux effected.

The internal date type of MAMA Date Time was changes to ensure it is
explicitly 64 bit where time_t is not big enough. This is to allow
extended datetime methods to be used even on 32 bit linux machines
(it already worked fine on 32 bit windows machines). It still
prefers time_t though and when support for 32 bit linux is
eventually dropped, the explicit value will be removed in favour
of the more OS-native type.

This issue is concerned with large year values (i.e. beyond the year
2038). Processing dates before 1970 is unlikely to cause a problem
(since that time_t range starts at 1902).

There are still limitations on 32 bit linux though. In order to get /
set an extended date time, the F64 accessors will need to be used:

    mamaDateTime_getEpochTimeSeconds
    mamaDateTime_setEpochTimeF64

The other methods will all work, but they may not all support extended
date time due to limitations which the size of time_t enforces in the
OS. In these instances, the effected methods will now return a
STATUS_INVALID_ARG, since the timestamp cannot be processed
correctly. This is mostly for string formatting functions and
functions for breaking strings out into their calendar components.

Signed-off-by: Frank Quinn <fquinn@velatt.com>