New extended epoch, hints and precision methods
The interface for getting and setting hints and precision was mixed up
with the non-extended methods for getting and setting the epoch time,
so this change introduces a few new methods for access to hints,
precision and epoch time with explicit 64 bit second types and
nanosecond resolution.

Relates to issues #285 and #286.

Signed-off-by: Frank Quinn <fquinn@velatt.com>